# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: security-research API contract
  version: 1.0.0
  description: |
    6 backend хувилбарт нэгдмэл API контракт.

    Судалгаа: "Орчин үеийн вэб фреймворкуудын өгөгдөл баталгаажуулалт болон ORM
    архитектурын халдлагаас хамгаалах чадварын харьцуулалт".

    Дүрэм: 6 backend бүгд ижил route, ижил статус код, ижил JSON хэлбэр гаргах ёстой.
    Хазайлт (jaltrolted) илэрвэл халдлагын detection буруу гарна.

    Дэлгэрэнгүй: `08-Дизайн-архитектур.md`, `04-Халдлагын-туршилт.md`, `11-Цель-өгөгдөл.md`

servers:
  - url: http://localhost:{port}
    description: Backend variant (variant branch дээр deploy хийхэд бодит URL болно)
    variables:
      port:
        default: '3001'
        enum: ['3001', '3002', '4001', '4002', '5001', '5002']

tags:
  - name: health
    description: Backend амьд эсэхийг шалгах
  - name: auth
    description: Хэрэглэгч бүртгэх, нэвтрэх
  - name: products
    description: Бараа жагсаалт, хайлт (SQLi туршилтын гол endpoint)
  - name: orders
    description: Захиалга (BOLA туршилтын гол endpoint)
  - name: profile
    description: Профайл шинэчлэх (JSON Overposting туршилтын endpoint)
  - name: admin
    description: Detection туршилтад ашиглах — халдлагын дараах target-уудын статус

paths:
  /health:
    get:
      tags: [health]
      summary: Backend healthcheck
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /api/auth/register:
    post:
      tags: [auth]
      summary: Шинэ хэрэглэгч бүртгэх (role=customer)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Хэрэглэгч үүсгэсэн
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: username/email аль хэдийн байна
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/login:
    post:
      tags: [auth]
      summary: Нэвтэрч JWT авах
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Нэвтэрсэн
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/ValidationError'

  /api/products:
    get:
      tags: [products]
      summary: Бараа жагсаалт (SQLi туршилтын entry point)
      description: |
        `search` параметр нь SQLi UNION/Boolean/Error 3 вектор туршилтын гол оролт.
        Raw хувилбарууд string interpolation-аар query бичих ёстой (vulnerable-by-design).
      parameters:
        - name: search
          in: query
          schema: { type: string }
          description: Барааны нэрэн доторх хайлт (LIKE)
        - name: category
          in: query
          schema: { type: integer }
        - name: min_price
          in: query
          schema: { type: number, format: float }
        - name: max_price
          in: query
          schema: { type: number, format: float }
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 50 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/products/{id}:
    get:
      tags: [products]
      summary: Нэг барааны дэлгэрэнгүй
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/orders:
    post:
      tags: [orders]
      summary: Захиалга үүсгэх
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Үүссэн
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/orders/{id}:
    get:
      tags: [orders]
      summary: Захиалга харах (BOLA GET туршилт)
      description: |
        Alpha implementation: зөвхөн `WHERE id=$1` — өөр хэрэглэгчийн захиалга гарна.
        Beta implementation: `WHERE id=$1 AND user_id=$2` — 404 буцаана.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [orders]
      summary: Захиалгын статус солих (BOLA PUT туршилт)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [orders]
      summary: Захиалга устгах (BOLA DELETE туршилт)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Устсан
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/profile:
    get:
      tags: [profile]
      summary: Өөрийн профайл
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags: [profile]
      summary: Профайл шинэчлэх (Overposting туршилт)
      description: |
        Alpha: DTO/Pydantic default `extra='ignore'` эсвэл `whitelist:false` —
        `targets[]` талбар силент accept болно.
        Beta: `whitelist:true, forbidNonWhitelisted:true` эсвэл `extra='forbid'` —
        400 буцаана.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/admin/targets/status:
    get:
      tags: [admin]
      summary: Халдлагын дараах target-уудын статус (detection)
      description: |
        Зөвхөн `attacks/detect_writes.py` болон pre_attack скриптүүд ашиглана.
        admin JWT шаардана. Халдлагын дараа write-target бүрд `target_value` эцсийн
        утгыг буцаана — R1 detection triple шалгах.
      parameters:
        - name: label
          in: query
          schema: { type: string }
          description: 'Тодорхой label шүүх (жишээ: WRITE_ORD_PUT_01)'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TargetsStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer }

  responses:
    ValidationError:
      description: Оролтын алдаа
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Хүчингүй JWT эсвэл нэвтэрч ороогүй
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Access denied (BOLA beta implementation-т)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Олдсонгүй
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Серверийн дотоод алдаа (stack trace илрэхгүй)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    HealthResponse:
      type: object
      required: [status, variant]
      properties:
        status: { type: string, enum: [ok] }
        variant:
          type: string
          description: 'Backend хувилбарын нэр (жишээ: nestjs-raw)'
          enum: [nestjs-raw, nestjs-orm, fiber-raw, fiber-orm, fastapi-raw, fastapi-orm]
        implementation:
          type: string
          enum: [alpha, beta]
          description: 'Alpha (default) эсвэл Beta (fixed) — feature flag-аар шилждэг'

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: 'Алдааны код (snake_case): validation_error | invalid_token | forbidden | not_found | internal | conflict'
        message:
          type: string
          description: 'Хэрэглэгчид харуулах мессеж (монголоор)'

    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username: { type: string, minLength: 3, maxLength: 50 }
        email:    { type: string, format: email }
        password: { type: string, minLength: 8 }

    RegisterResponse:
      type: object
      required: [user_id, username, role]
      properties:
        user_id:  { type: integer }
        username: { type: string }
        role:     { type: string, enum: [customer] }

    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }

    LoginResponse:
      type: object
      required: [access_token, role, user_id]
      properties:
        access_token: { type: string, description: 'JWT HS256, exp=15m' }
        role:         { type: string, enum: [customer, admin] }
        user_id:      { type: integer }

    Product:
      type: object
      required: [id, name, price]
      properties:
        id:          { type: integer }
        name:        { type: string }
        description: { type: string, nullable: true }
        price:       { type: number, format: float }
        category_id: { type: integer, nullable: true }

    ProductListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Product' }

    ProductResponse:
      type: object
      required: [product]
      properties:
        product: { $ref: '#/components/schemas/Product' }

    OrderItem:
      type: object
      required: [product_id, quantity, unit_price]
      properties:
        product_id: { type: integer }
        quantity:   { type: integer, minimum: 1 }
        unit_price: { type: number, format: float }

    Order:
      type: object
      required: [id, user_id, status, total, items]
      properties:
        id:      { type: integer }
        user_id: { type: integer }
        status:  { type: string, enum: [pending, paid, shipped, cancelled] }
        total:   { type: number, format: float }
        items:
          type: array
          items: { $ref: '#/components/schemas/OrderItem' }

    CreateOrderRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [product_id, quantity]
            properties:
              product_id: { type: integer }
              quantity:   { type: integer, minimum: 1 }

    UpdateOrderRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          description: |
            Хүчинтэй утгууд: pending | paid | shipped | cancelled.
            Alpha implementation-т enum шалгагдахгүй — халдлагын `"You are hacked"`
            гэсэн утгыг backend `order_targets.target_value`-т `"You are hacked | <nonce>"`
            хэлбэрээр бичдэг.

    OrderResponse:
      type: object
      required: [order]
      properties:
        order: { $ref: '#/components/schemas/Order' }

    Profile:
      type: object
      required: [user_id, username, email, role, is_admin]
      properties:
        user_id:  { type: integer }
        username: { type: string }
        email:    { type: string, format: email }
        role:     { type: string, enum: [customer, admin] }
        is_admin: { type: boolean }
        address:  { type: string, nullable: true }

    UpdateProfileRequest:
      type: object
      description: |
        Alpha implementation нь мэдэгдээгүй талбарыг (`targets`, `role`, `is_admin`)
        силент accept хийж болно — энэ бол Overposting туршилтын гол цэг.
      properties:
        name:    { type: string }
        address: { type: string }
        # Дараах талбарууд нь Overposting халдлагын оролт:
        role:    { type: string, description: 'Overposting variant 1 — role escalation' }
        is_admin: { type: boolean, description: 'Overposting variant 2 — admin escalation' }
        targets:
          type: array
          description: 'Overposting variant 3 — profile_targets-т шууд бичих оролдлого'
          items:
            type: object
            required: [label, value]
            properties:
              label: { type: string }
              value: { type: string }

    ProfileResponse:
      type: object
      required: [profile]
      properties:
        profile: { $ref: '#/components/schemas/Profile' }

    TargetsStatusResponse:
      type: object
      required: [targets]
      properties:
        targets:
          type: array
          items:
            type: object
            required: [label, kind, value, nonce]
            properties:
              label:
                type: string
                description: 'READ_UNION_01 | READ_BOOL_01 | READ_ERR_01 | WRITE_ORD_PUT_01 | WRITE_ORD_DEL_01 | WRITE_PROF_01'
              kind:
                type: string
                enum: [read, write]
              value:
                type: string
                description: |
                  Read: `secrets.secret_value` (жишээ нь "you got right data").
                  Write: `order_targets.target_value` эсвэл `profile_targets.target_value`
                  (одоогийн утга — success/blocked detection-т ашиглана).
              nonce:
                type: string
              deleted:
                type: boolean
                description: 'BOLA DELETE-д: тухайн order устсан эсэх'
                default: false

security:
  - bearerAuth: []